Fix duplicate Bluesky links
If a person has only a Bluesky link, it was displayed twice. It
looks like support was added in two independent PRs: #1872, and
#1878.

You can see an example of the issue on the single profile
displayed here: https://uses.tech/like/Simplicity

I've also tweaked how the handles are normalised. Previously one
block removed '@'s in the URL, then blindly chomped off the
first character in the display name, while the second assumed
there was no `@` prefix. There are a few entries that have an '@',
so I've stuck with just removing it, as it'll work in all current
cases.